Improvements of the consoleReader extractor.
The resulting extractor is now of type SingleOptionValue[String]
because only a single value is read from the console.

It is now possible to specify a prompt string, which is used instead
of the option key if defined.

The password flag is now true per default because reading passwords is
the typical use case for this function.