Added a new implementation to parse command line parameters.
This implementation is based on a classifier function, which expects a
command line element and returns its type. Based on the type, the
element is then processed. This new implementation supports switches.

As the old implementation is still referenced by other classes, it has
been kept for now (but renamed to Old); it is going to be removed
later on.