A shortAlias flag can now be passed to options and switches.
When constructing the CliExtractor for an option or a switch, it is
now possible to state whether the key provided is a full key or a
short alias. The default is that the keys passed in here are full
keys, while aliases can be declared later; but users can change this.

Renamed references to ModelContext in CliExtractorHelpSpec.