Implemented support for aliases in ParameterManager.
ParameterManager now provides a function to resolve aliases based on a
(now bidirectional) AliasMapping.

The API of ParameterManager has been slightly changed to make sure
that the current CliExtractor is executed at most once to gather
metadata. To achieve this, a new ExtractorContext class has been
introduced which constructs a ParameterContext from the extractor
lazily when it is needed.