Added a new function for processing parameter files.
Options referencing files with parameters to be included are now
handled by a separate function. This function accepts the original
sequence of command line arguments, searches for files to be included,
and yields the augmented command line.

The regular parse() function is going to be reworked to no longer
handle parameter files.