Reworked ExtractorContext class.
The class is now named ExtractionSpec and defines a number of settings
to customize the extraction operation. This simplifies the API of
ParameterManager, as the options no longer need to be passed to
functions.

ExtractionSpec still provides access to the metadata of the wrapped
CliExtractor. However, it now makes sure that the options referencing
parameter files are included in the ModelContext. This is necessary to
make sure that they are correctly classified. (When combined switches
are enabled, short aliases for file options were misinterpreted as
switches.)