Fixed TODOs related to correct keys for input parameters.
If a key for the parameter affected is specified, this key is now used
for exceptions. There is still a new TODO because there must be a way
to distinguish such keys from normal keys; so that a renderer knows
that no prefix needs to be added.

Renamed the reserved key for input parameters from InputOption to
InputParameter.