ParameterManager now supports help flags when processing parameters.
In the ExtractionSpec, an application can now specify a special
CliExtractor to detect a help flag on the command line, such as --help
or -h. The result of the processCommandLine() methods has been changed
to report the result of this help extractor. Hence, a caller can
easily find out whether the user has requested help.

Having a help extractor separate from the actual extractor used by the
application avoids mixing the parameter representation of the
application with technical aspects.