Fixed a problem with duplicate aliases.
If an extractor defining aliases was invoked multiple times (e.g.
during execution of conditional logic), the aliases were duplicated in
the model context. So they were also displayed multiple times in help
texts.

This has been fixed; a new alias is now added to the alias mapping
only if it is not already contained in it.