Introduced ParameterMetaDataSource trait.
This trait provides access to a collection of ParameterMetaData
objects. It is extended by ModelContext, which manages the metadata of
all the parameters known to the system. HelpGenerator now expects a
ParameterMetaDataSource rather than a ModelContext when generating
help tables.

The goal is to add another context class extending this trait for
storing error information. The functionality of HelpGenerator can then
be used analogously for the new error context.