Introduced ParameterFailure trait.
The trait represents a failure that was detected during parameter
processing. It is implemented by the ExtractionFailure class.

The main reason for this trait is that it is going to be used by the
ParameterModel module, but ExtractionFailure is contained in
ParameterExtractor; ParameterModel must not access ParameterExtractor
because the dependency is in the other direction. In addition, not all
the information contained in ExtractionFailure is actually required.