The extractor for single values now allows overriding values.
In this mode, the extractor accepts multiple values specified on the
command last; the last value is selected. This is especially helpful
when working with parameter files: The file may already contain a
default value for a parameter, but it is possible to override this
parameter on the command line.