excluding() now allows overriding values.
In this mode, multiple options are allowed to have values. The option
that occurs on the latest position in the command line wins. To
achieve this, the excluding extractors that yield a value are sorted
by the positions reported by the CliElements defining their values.
This only works if the extractors have direct representations on the
command line.